Network.AWS.DynamoDBStreams

Description

Amazon DynamoDB

Amazon DynamoDB Streams provides API actions for accessing streams and processing stream records. To learn more about application development with Streams, see Capturing Table Activity with DynamoDB Streams in the Amazon DynamoDB Developer Guide.

Service Configuration

dynamoDBStreams :: Service

API version 2012-08-10 of the Amazon DynamoDB Streams SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by DynamoDBStreams.

ExpiredIteratorException

_ExpiredIteratorException :: AsError a => Getting (First ServiceError) a ServiceError

The shard iterator has expired and can no longer be used to retrieve stream records. A shard iterator expires 15 minutes after it is retrieved using the GetShardIterator action.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred on the server side.

TrimmedDataAccessException

_TrimmedDataAccessException :: AsError a => Getting (First ServiceError) a ServiceError

The operation attempted to read past the oldest stream record in a shard.

In DynamoDB Streams, there is a 24 hour limit on data retention. Stream records whose age exceeds this limit are subject to removal (trimming) from the stream. You might receive a TrimmedDataAccessException if:

You request a shard iterator with a sequence number older than the trim point (24 hours).

You obtain a shard iterator, but before you use the iterator in a GetRecords request, a stream record in the shard exceeds the 24 hour period and is trimmed. This causes the iterator to access a record that no longer exists.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The operation tried to access a nonexistent stream.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your request rate is too high. The AWS SDKs for DynamoDB automatically retry requests that receive this exception. Your request is eventually successful, unless your retry queue is too large to finish. Reduce the frequency of requests and use exponential backoff. For more information, go to Error Retries and Exponential Backoff in the Amazon DynamoDB Developer Guide .

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetShardIterator

module Network.AWS.DynamoDBStreams.GetShardIterator

GetRecords

module Network.AWS.DynamoDBStreams.GetRecords

ListStreams

module Network.AWS.DynamoDBStreams.ListStreams

DescribeStream

module Network.AWS.DynamoDBStreams.DescribeStream

Types

KeyType

data KeyType

Constructors

Hash

Range

OperationType

data OperationType

Constructors

Insert

Modify

Remove

ShardIteratorType

data ShardIteratorType

Constructors

AfterSequenceNumber

AtSequenceNumber

Latest

TrimHorizon

StreamStatus

data StreamStatus

Constructors

Disabled

Disabling

Enabled

Enabling

StreamViewType

data StreamViewType

Constructors

KeysOnly

NewAndOldImages

NewImage

OldImage

AttributeValue

data AttributeValue

Represents the data for an attribute. You can set one, and only one, of the elements.

Each attribute in an item is a name-value pair. An attribute can be single-valued or multi-valued set. For example, a book item can have title and authors attributes. Each book has one title but can have many authors. The multi-valued attribute is a set; duplicate values are not allowed.

See: attributeValue smart constructor.

attributeValue :: AttributeValue

Creates a value of AttributeValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avL - A List data type.

avNS - A Number Set data type.

avM - A Map data type.

avNULL - A Null data type.

avN - A Number data type.

avBS - A Binary Set data type.

avB - A Binary data type.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

avSS - A String Set data type.

avS - A String data type.

avBOOL - A Boolean data type.

avL :: Lens' AttributeValue [AttributeValue]

A List data type.

avNS :: Lens' AttributeValue [Text]

A Number Set data type.

avM :: Lens' AttributeValue (HashMap Text AttributeValue)

A Map data type.

avNULL :: Lens' AttributeValue (Maybe Bool)

A Null data type.

avN :: Lens' AttributeValue (Maybe Text)

A Number data type.

avBS :: Lens' AttributeValue [ByteString]

A Binary Set data type.

avB :: Lens' AttributeValue (Maybe ByteString)

A Binary data type.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

avSS :: Lens' AttributeValue [Text]

A String Set data type.

avS :: Lens' AttributeValue (Maybe Text)

A String data type.

avBOOL :: Lens' AttributeValue (Maybe Bool)

A Boolean data type.

Identity

data Identity

Contains details about the type of identity that made the request.

See: identity smart constructor.

identity :: Identity

Creates a value of Identity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iPrincipalId - A unique identifier for the entity that made the call. For Time To Live, the principalId is "dynamodb.amazonaws.com".

iType - The type of the identity. For Time To Live, the type is Service.

iPrincipalId :: Lens' Identity (Maybe Text)

A unique identifier for the entity that made the call. For Time To Live, the principalId is "dynamodb.amazonaws.com".

iType :: Lens' Identity (Maybe Text)

The type of the identity. For Time To Live, the type is Service.

KeySchemaElement

data KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes that make up the primary key of a table, or the key attributes of an index.

A KeySchemaElement represents exactly one attribute of the primary key. For example, a simple primary key (partition key) would be represented by one KeySchemaElement . A composite primary key (partition key and sort key) would require one KeySchemaElement for the partition key, and another KeySchemaElement for the sort key.

See: keySchemaElement smart constructor.

keySchemaElement

Arguments

:: Text
kseAttributeName

-> KeyType
kseKeyType

-> KeySchemaElement

Creates a value of KeySchemaElement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kseAttributeName - The name of a key attribute.

kseKeyType - The attribute data, consisting of the data type and the attribute value itself.

kseAttributeName :: Lens' KeySchemaElement Text

The name of a key attribute.

kseKeyType :: Lens' KeySchemaElement KeyType

The attribute data, consisting of the data type and the attribute value itself.

Record

data Record

A description of a unique event within a stream.

See: record smart constructor.

record :: Record

Creates a value of Record with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rUserIdentity - Items that are deleted by the Time to Live process after expiration have the following fields: * Records[].userIdentity.type Service * Records[].userIdentity.principalId "dynamodb.amazonaws.com"

rEventVersion - The version number of the stream record format. This number is updated whenever the structure of Record is modified. Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

rDynamodb - The main body of the stream record, containing all of the DynamoDB-specific fields.

rAwsRegion - The region in which the GetRecords request was received.

rEventName - The type of data modification that was performed on the DynamoDB table: * INSERT - a new item was added to the table. * MODIFY - one or more of an existing item's attributes were modified. * REMOVE - the item was deleted from the table

rEventSource - The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb .

rEventId - A globally unique identifier for the event that was recorded in this stream record.

rUserIdentity :: Lens' Record (Maybe Identity)

Items that are deleted by the Time to Live process after expiration have the following fields: * Records[].userIdentity.type Service * Records[].userIdentity.principalId "dynamodb.amazonaws.com"

rEventVersion :: Lens' Record (Maybe Text)

The version number of the stream record format. This number is updated whenever the structure of Record is modified. Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

rDynamodb :: Lens' Record (Maybe StreamRecord)

The main body of the stream record, containing all of the DynamoDB-specific fields.

rAwsRegion :: Lens' Record (Maybe Text)

The region in which the GetRecords request was received.

rEventName :: Lens' Record (Maybe OperationType)

The type of data modification that was performed on the DynamoDB table: * INSERT - a new item was added to the table. * MODIFY - one or more of an existing item's attributes were modified. * REMOVE - the item was deleted from the table

rEventSource :: Lens' Record (Maybe Text)

The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb .

rEventId :: Lens' Record (Maybe Text)

A globally unique identifier for the event that was recorded in this stream record.

SequenceNumberRange

data SequenceNumberRange

The beginning and ending sequence numbers for the stream records contained within a shard.

See: sequenceNumberRange smart constructor.

sequenceNumberRange :: SequenceNumberRange

Creates a value of SequenceNumberRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

snrStartingSequenceNumber - The first sequence number.

snrEndingSequenceNumber - The last sequence number.

snrStartingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The first sequence number.

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The last sequence number.

Shard

data Shard

A uniquely identified group of stream records within a stream.

See: shard smart constructor.

shard :: Shard

Creates a value of Shard with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sParentShardId - The shard ID of the current shard's parent.

sSequenceNumberRange - The range of possible sequence numbers for the shard.

sShardId - The system-generated identifier for this shard.

sParentShardId :: Lens' Shard (Maybe Text)

The shard ID of the current shard's parent.

sSequenceNumberRange :: Lens' Shard (Maybe SequenceNumberRange)

The range of possible sequence numbers for the shard.

sShardId :: Lens' Shard (Maybe Text)

The system-generated identifier for this shard.

Stream

data Stream

Represents all of the data describing a particular stream.

See: stream smart constructor.

stream :: Stream

Creates a value of Stream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sStreamLabel - A timestamp, in ISO 8601 format, for this stream. Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique: * the AWS customer ID. * the table name * the StreamLabel

sStreamARN - The Amazon Resource Name (ARN) for the stream.

sTableName - The DynamoDB table with which the stream is associated.

sStreamLabel :: Lens' Stream (Maybe Text)

A timestamp, in ISO 8601 format, for this stream. Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique: * the AWS customer ID. * the table name * the StreamLabel

sStreamARN :: Lens' Stream (Maybe Text)

The Amazon Resource Name (ARN) for the stream.

sTableName :: Lens' Stream (Maybe Text)

The DynamoDB table with which the stream is associated.

StreamDescription

data StreamDescription

Represents all of the data describing a particular stream.

See: streamDescription smart constructor.

streamDescription :: StreamDescription

Creates a value of StreamDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdLastEvaluatedShardId - The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request. If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved. If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

sdStreamLabel - A timestamp, in ISO 8601 format, for this stream. Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique: * the AWS customer ID. * the table name * the StreamLabel

sdStreamStatus - Indicates the current status of the stream: * ENABLING - Streams is currently being enabled on the DynamoDB table. * ENABLED - the stream is enabled. * DISABLING - Streams is currently being disabled on the DynamoDB table. * DISABLED - the stream is disabled.

sdKeySchema - The key attribute(s) of the stream's DynamoDB table.

sdStreamViewType - Indicates the format of the records within this stream: * KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table. * NEW_IMAGE - entire items from the table, as they appeared after they were modified. * OLD_IMAGE - entire items from the table, as they appeared before they were modified. * NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

sdStreamARN - The Amazon Resource Name (ARN) for the stream.

sdShards - The shards that comprise the stream.

sdTableName - The DynamoDB table with which the stream is associated.

sdCreationRequestDateTime - The date and time when the request to create this stream was issued.

sdLastEvaluatedShardId :: Lens' StreamDescription (Maybe Text)

The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request. If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved. If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

sdStreamLabel :: Lens' StreamDescription (Maybe Text)

A timestamp, in ISO 8601 format, for this stream. Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique: * the AWS customer ID. * the table name * the StreamLabel

sdStreamStatus :: Lens' StreamDescription (Maybe StreamStatus)

Indicates the current status of the stream: * ENABLING - Streams is currently being enabled on the DynamoDB table. * ENABLED - the stream is enabled. * DISABLING - Streams is currently being disabled on the DynamoDB table. * DISABLED - the stream is disabled.

sdKeySchema :: Lens' StreamDescription (Maybe (NonEmpty KeySchemaElement))

The key attribute(s) of the stream's DynamoDB table.

sdStreamViewType :: Lens' StreamDescription (Maybe StreamViewType)

Indicates the format of the records within this stream: * KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table. * NEW_IMAGE - entire items from the table, as they appeared after they were modified. * OLD_IMAGE - entire items from the table, as they appeared before they were modified. * NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

sdStreamARN :: Lens' StreamDescription (Maybe Text)

The Amazon Resource Name (ARN) for the stream.

sdShards :: Lens' StreamDescription [Shard]

The shards that comprise the stream.

sdTableName :: Lens' StreamDescription (Maybe Text)

The DynamoDB table with which the stream is associated.

sdCreationRequestDateTime :: Lens' StreamDescription (Maybe UTCTime)

The date and time when the request to create this stream was issued.

StreamRecord

data StreamRecord

A description of a single data modification that was performed on an item in a DynamoDB table.

See: streamRecord smart constructor.

streamRecord :: StreamRecord

Creates a value of StreamRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srSizeBytes - The size of the stream record, in bytes.

srSequenceNumber - The sequence number of the stream record.

srApproximateCreationDateTime - The approximate date and time when the stream record was created, in UNIX epoch time format.

srStreamViewType - The type of data from the modified DynamoDB item that was captured in this stream record: * KEYS_ONLY - only the key attributes of the modified item. * NEW_IMAGE - the entire item, as it appeared after it was modified. * OLD_IMAGE - the entire item, as it appeared before it was modified. * NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

srKeys - The primary key attribute(s) for the DynamoDB item that was modified.

srOldImage - The item in the DynamoDB table as it appeared before it was modified.

srNewImage - The item in the DynamoDB table as it appeared after it was modified.

srSizeBytes :: Lens' StreamRecord (Maybe Natural)

The size of the stream record, in bytes.

srSequenceNumber :: Lens' StreamRecord (Maybe Text)

The sequence number of the stream record.

srApproximateCreationDateTime :: Lens' StreamRecord (Maybe UTCTime)

The approximate date and time when the stream record was created, in UNIX epoch time format.

srStreamViewType :: Lens' StreamRecord (Maybe StreamViewType)

The type of data from the modified DynamoDB item that was captured in this stream record: * KEYS_ONLY - only the key attributes of the modified item. * NEW_IMAGE - the entire item, as it appeared after it was modified. * OLD_IMAGE - the entire item, as it appeared before it was modified. * NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

srKeys :: Lens' StreamRecord (HashMap Text AttributeValue)

The primary key attribute(s) for the DynamoDB item that was modified.

srOldImage :: Lens' StreamRecord (HashMap Text AttributeValue)

The item in the DynamoDB table as it appeared before it was modified.

srNewImage :: Lens' StreamRecord (HashMap Text AttributeValue)

The item in the DynamoDB table as it appeared after it was modified.