Network.AWS.DynamoDBStreams.ListStreams

Description

Returns an array of stream ARNs associated with the current account and endpoint. If the TableName parameter is present, then ListStreams will return only the streams ARNs for that table.

Creating a Request

listStreams :: ListStreams

Creates a value of ListStreams with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsExclusiveStartStreamARN - The ARN (Amazon Resource Name) of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedStreamArn in the previous operation.

lsLimit - The maximum number of streams to return. The upper limit is 100.

lsTableName - If this parameter is provided, then only the streams associated with this table name are returned.

data ListStreams

Represents the input of a ListStreams operation.

See: listStreams smart constructor.

Request Lenses

lsExclusiveStartStreamARN :: Lens' ListStreams (Maybe Text)

The ARN (Amazon Resource Name) of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedStreamArn in the previous operation.

lsLimit :: Lens' ListStreams (Maybe Natural)

The maximum number of streams to return. The upper limit is 100.

lsTableName :: Lens' ListStreams (Maybe Text)

If this parameter is provided, then only the streams associated with this table name are returned.

Destructuring the Response

listStreamsResponse

Arguments

:: Int
lsrsResponseStatus

-> ListStreamsResponse

Creates a value of ListStreamsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsrsLastEvaluatedStreamARN - The stream ARN of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request. If LastEvaluatedStreamArn is empty, then the "last page" of results has been processed and there is no more data to be retrieved. If LastEvaluatedStreamArn is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedStreamArn is empty.

lsrsStreams - A list of stream descriptors associated with the current account and endpoint.

lsrsResponseStatus - -- | The response status code.

data ListStreamsResponse

Represents the output of a ListStreams operation.

See: listStreamsResponse smart constructor.

Response Lenses

lsrsLastEvaluatedStreamARN :: Lens' ListStreamsResponse (Maybe Text)

The stream ARN of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request. If LastEvaluatedStreamArn is empty, then the "last page" of results has been processed and there is no more data to be retrieved. If LastEvaluatedStreamArn is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedStreamArn is empty.

lsrsStreams :: Lens' ListStreamsResponse [Stream]

A list of stream descriptors associated with the current account and endpoint.

lsrsResponseStatus :: Lens' ListStreamsResponse Int

- | The response status code.